requirements.txt
certifi==2021.5.30
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==2.0.1
marshmallow==3.17.0
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.40
SQLAlchemy-ImageAttach==1.1.0
Wand==0.6.9
Werkzeug==1.0.1
wincertstore==0.2
